feat: Implement Huffman algorithm for file compression
This commit adds the implementation of the Huffman algorithm for file compression. The code changes include updating the compressor.cpp file to include the necessary headers and libraries, as well as adding the logic to encode the data using the Huffman codes. The compressed data is then written to a binary file with the extension ".huff". The commit also includes recent user commits that mention the addition of the Huffman algorithm.